requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
markdown==3.7
Pygments==2.18.0
Pillow==11.1.0
bleach==6.2.0
python-slugify==8.0.4
python-dotenv==1.0.1
PyYAML==6.0.2
gunicorn==23.0.0
